realizando update de usuario

+++ src/main/java/Zeus/API/ZEUS/Dto/DadosAtualizacaoRacao.java
@@ -11,7 +11,7 @@ public record DadosAtualizacaoRacao(
         Long id,
         String nome,
         int kgQuantidade,
-        double valorPago,
+        BigDecimal valorPago,
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd/MM/yyyy")
         LocalDate dataCompra
